Refactor person Analizer. Extended added.
